separating data into two separate files
