+++ src/data/socials.json
@@ -1,0 +1,14 @@
+{
+    "socials": [
+      {
+        "name": "Github",
+        "description": "Looking to contribute?",
+        "url": "https://github.com/bhelpful"
+      },
+      {
+        "name": "LinkedIn",
+        "description": "Connect with us",
+        "url": "https://linkedin.com/company/bhelpful"
+      }
+    ]
+}
